tenant parameter removed
