+++ src/Weaviate.Client/GenerateClient.NearText.cs
@@ -1,0 +1,147 @@
+using Weaviate.Client.Models;
+
+namespace Weaviate.Client;
+
+public partial class GenerateClient
+{
+    /// <summary>
+    /// Search near text with generative AI capabilities.
+    /// </summary>
+    /// <param name="text">Text to search near</param>
+    /// <param name="certainty">Certainty threshold</param>
+    /// <param name="distance">Distance threshold</param>
+    /// <param name="moveTo">Move towards concept</param>
+    /// <param name="moveAway">Move away from concept</param>
+    /// <param name="limit">Maximum number of results</param>
+    /// <param name="offset">Offset for pagination</param>
+    /// <param name="autoLimit">Auto-cut threshold</param>
+    /// <param name="filters">Filters to apply</param>
+    /// <param name="rerank">Rerank configuration</param>
+    /// <param name="prompt">Single prompt for generation</param>
+    /// <param name="groupedTask">Grouped prompt for generation</param>
+    /// <param name="targetVector">Target vector name</param>
+    /// <param name="returnProperties">Properties to return</param>
+    /// <param name="returnReferences">References to return</param>
+    /// <param name="returnMetadata">Metadata to return</param>
+    /// <param name="includeVectors">Vectors to include</param>
+    /// <param name="cancellationToken">Cancellation token for the operation</param>
+    /// <returns>Generative result</returns>
+    public async Task<GenerativeWeaviateResult> NearText(
+        OneOrManyOf<string> text,
+        float? certainty = null,
+        float? distance = null,
+        Move? moveTo = null,
+        Move? moveAway = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        SinglePrompt? prompt = null,
+        GroupedTask? groupedTask = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    )
+    {
+        var result = await _client.GrpcClient.SearchNearText(
+            _collectionClient.Name,
+            text.ToArray(),
+            distance: distance,
+            certainty: certainty,
+            limit: limit,
+            moveTo: moveTo,
+            moveAway: moveAway,
+            offset: offset,
+            autoLimit: autoLimit,
+            targetVector: targetVector,
+            filters: filters,
+            tenant: _collectionClient.Tenant,
+            rerank: rerank,
+            singlePrompt: prompt,
+            groupedTask: groupedTask,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+        return result;
+    }
+
+    /// <summary>
+    /// Search near text with generative AI capabilities and grouping.
+    /// </summary>
+    /// <param name="text">Text to search near</param>
+    /// <param name="groupBy">Group by configuration</param>
+    /// <param name="certainty">Certainty threshold</param>
+    /// <param name="distance">Distance threshold</param>
+    /// <param name="moveTo">Move towards concept</param>
+    /// <param name="moveAway">Move away from concept</param>
+    /// <param name="limit">Maximum number of results</param>
+    /// <param name="offset">Offset for pagination</param>
+    /// <param name="autoLimit">Auto-cut threshold</param>
+    /// <param name="filters">Filters to apply</param>
+    /// <param name="rerank">Rerank configuration</param>
+    /// <param name="prompt">Single prompt for generation</param>
+    /// <param name="groupedTask">Grouped prompt for generation</param>
+    /// <param name="targetVector">Target vector name</param>
+    /// <param name="returnProperties">Properties to return</param>
+    /// <param name="returnReferences">References to return</param>
+    /// <param name="returnMetadata">Metadata to return</param>
+    /// <param name="includeVectors">Vectors to include</param>
+    /// <param name="cancellationToken">Cancellation token for the operation</param>
+    /// <returns>Generative group-by result</returns>
+    public async Task<GenerativeGroupByResult> NearText(
+        OneOrManyOf<string> text,
+        GroupByRequest groupBy,
+        float? certainty = null,
+        float? distance = null,
+        Move? moveTo = null,
+        Move? moveAway = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        SinglePrompt? prompt = null,
+        GroupedTask? groupedTask = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    )
+    {
+        var result = await _client.GrpcClient.SearchNearText(
+            _collectionClient.Name,
+            text.ToArray(),
+            groupBy: groupBy,
+            distance: distance,
+            certainty: certainty,
+            moveTo: moveTo,
+            moveAway: moveAway,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            tenant: _collectionClient.Tenant,
+            rerank: rerank,
+            singlePrompt: prompt,
+            groupedTask: groupedTask,
+            targetVector: targetVector,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+        return result;
+    }
+}
