+++ src/Weaviate.Client/GenerateClient.cs
@@ -7,7 +7,7 @@ public partial class CollectionClient
     public GenerateClient Generate => new(this);
 }
 
-public class GenerateClient
+public partial class GenerateClient
 {
     private readonly CollectionClient _collectionClient;
     private string _collectionName => _collectionClient.Name;
@@ -30,1421 +30,4 @@ public class GenerateClient
             _client.QueryTimeout ?? _client.DefaultTimeout ?? WeaviateDefaults.QueryTimeout;
         return TimeoutHelper.GetCancellationToken(effectiveTimeout, userToken);
     }
-
-    #region Objects
-    /// <summary>
-    /// Fetch objects with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="sort">Sort configuration</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> FetchObjects(
-        Models.GroupByRequest groupBy,
-        Guid? after = null,
-        uint? limit = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _client.GrpcClient.FetchObjects(
-            _collectionName,
-            after: after,
-            limit: limit,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            filters: filters,
-            sort: sort,
-            groupBy: groupBy,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    /// <summary>
-    /// Fetch objects with generative AI capabilities.
-    /// </summary>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="sort">Sort configuration</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult?> FetchObjects(
-        Guid? after = null,
-        uint? limit = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _client.GrpcClient.FetchObjects(
-            _collectionName,
-            after: after,
-            limit: limit,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            filters: filters,
-            sort: sort,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    /// <summary>
-    /// Fetch a single object by ID with generative AI capabilities.
-    /// </summary>
-    /// <param name="id">Object ID</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult?> FetchObjectByID(
-        Guid id,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.FetchObjects(
-            _collectionName,
-            returnProperties: returnProperties,
-            filters: Filter.WithID(id),
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            tenant: tenant ?? _collectionClient.Tenant,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// Fetch multiple objects by IDs with generative AI capabilities.
-    /// </summary>
-    /// <param name="ids">Set of object IDs</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="sort">Sort configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> FetchObjectsByIDs(
-        HashSet<Guid> ids,
-        uint? limit = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        if (ids == null)
-            throw new ArgumentNullException(nameof(ids));
-
-        if (ids.Count == 0)
-            return GenerativeWeaviateResult.Empty;
-
-        Filter idFilter = ids.Count == 1 ? Filter.WithID(ids.First()) : Filter.WithIDs(ids);
-
-        if (filters is not null)
-            idFilter = filters & idFilter;
-
-        return await _client.GrpcClient.FetchObjects(
-            _collectionName,
-            limit: limit,
-            filters: idFilter,
-            sort: sort,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-    }
-    #endregion
-
-    #region Search
-
-    /// <summary>
-    /// Search near text with generative AI capabilities.
-    /// </summary>
-    /// <param name="text">Text to search near</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="moveTo">Move towards concept</param>
-    /// <param name="moveAway">Move away from concept</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> NearText(
-        OneOrManyOf<string> text,
-        float? certainty = null,
-        float? distance = null,
-        Move? moveTo = null,
-        Move? moveAway = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearText(
-            _collectionClient.Name,
-            text.ToArray(),
-            distance: distance,
-            certainty: certainty,
-            limit: limit,
-            moveTo: moveTo,
-            moveAway: moveAway,
-            offset: offset,
-            autoLimit: autoLimit,
-            targetVector: targetVector,
-            filters: filters,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// Search near text with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="text">Text to search near</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="moveTo">Move towards concept</param>
-    /// <param name="moveAway">Move away from concept</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> NearText(
-        OneOrManyOf<string> text,
-        GroupByRequest groupBy,
-        float? certainty = null,
-        float? distance = null,
-        Move? moveTo = null,
-        Move? moveAway = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearText(
-            _collectionClient.Name,
-            text.ToArray(),
-            groupBy: groupBy,
-            distance: distance,
-            certainty: certainty,
-            moveTo: moveTo,
-            moveAway: moveAway,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            tenant: _collectionClient.Tenant,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// Search near vector with generative AI capabilities.
-    /// </summary>
-    /// <param name="vector">Vector to search near</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> NearVector(
-        Vectors vector,
-        Filter? filters = null,
-        float? certainty = null,
-        float? distance = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearVector(
-            _collectionClient.Name,
-            vector,
-            distance: distance,
-            certainty: certainty,
-            offset: offset,
-            autoLimit: autoLimit,
-            limit: limit,
-            targetVector: targetVector,
-            filters: filters,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// Search near vector with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="vector">Vector to search near</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> NearVector(
-        Vectors vector,
-        GroupByRequest groupBy,
-        Filter? filters = null,
-        float? distance = null,
-        float? certainty = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearVector(
-            _collectionClient.Name,
-            vector,
-            groupBy,
-            filters: filters,
-            distance: distance,
-            certainty: certainty,
-            offset: offset,
-            autoLimit: autoLimit,
-            limit: limit,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// BM25 search with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="searchFields">Fields to search in</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="after">Cursor for pagination</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="consistencyLevel">Consistency level</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> BM25(
-        string query,
-        GroupByRequest groupBy,
-        string[]? searchFields = null,
-        Filter? filters = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        Guid? after = null,
-        string? tenant = null,
-        ConsistencyLevels? consistencyLevel = null,
-        OneOrManyOf<string>? returnProperties = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        IList<QueryReference>? returnReferences = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchBM25(
-            _collectionClient.Name,
-            query: query,
-            searchFields: searchFields,
-            filters: filters,
-            autoLimit: autoLimit,
-            limit: limit,
-            offset: offset,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            after: after,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: consistencyLevel ?? _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnReferences: returnReferences,
-            returnProperties: returnProperties,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-        return result;
-    }
-
-    /// <summary>
-    /// BM25 search with generative AI capabilities.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="searchFields">Fields to search in</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="autoLimit">Auto-cut threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="after">Cursor for pagination</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="consistencyLevel">Consistency level</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> BM25(
-        string query,
-        string[]? searchFields = null,
-        Filter? filters = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        Guid? after = null,
-        string? tenant = null,
-        ConsistencyLevels? consistencyLevel = null,
-        OneOrManyOf<string>? returnProperties = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        IList<QueryReference>? returnReferences = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchBM25(
-            _collectionClient.Name,
-            query: query,
-            searchFields: searchFields,
-            filters: filters,
-            autoLimit: autoLimit,
-            limit: limit,
-            offset: offset,
-            groupBy: null,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            after: after,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: consistencyLevel ?? _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnReferences: returnReferences,
-            returnProperties: returnProperties,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Hybrid search with generative AI capabilities.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="alpha">Alpha value for hybrid search</param>
-    /// <param name="queryProperties">Properties to query</param>
-    /// <param name="fusionType">Fusion type</param>
-    /// <param name="maxVectorDistance">Maximum vector distance</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="bm25Operator">BM25 operator</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> Hybrid(
-        string? query,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Hybrid search with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="alpha">Alpha value for hybrid search</param>
-    /// <param name="queryProperties">Properties to query</param>
-    /// <param name="fusionType">Fusion type</param>
-    /// <param name="maxVectorDistance">Maximum vector distance</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="bm25Operator">BM25 operator</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> Hybrid(
-        string? query,
-        Models.GroupByRequest groupBy,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Hybrid search with generative AI capabilities using vectors.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="vectors">Vectors for search</param>
-    /// <param name="alpha">Alpha value for hybrid search</param>
-    /// <param name="queryProperties">Properties to query</param>
-    /// <param name="fusionType">Fusion type</param>
-    /// <param name="maxVectorDistance">Maximum vector distance</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="bm25Operator">BM25 operator</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> Hybrid(
-        string? query,
-        Vectors vectors,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await Hybrid(
-            query,
-            vectors: vectors,
-            alpha,
-            queryProperties,
-            fusionType,
-            maxVectorDistance,
-            limit,
-            offset,
-            bm25Operator,
-            autoLimit,
-            filters,
-            rerank,
-            prompt,
-            groupedTask,
-            targetVector,
-            tenant,
-            returnProperties,
-            returnReferences,
-            returnMetadata,
-            includeVectors,
-            cancellationToken
-        );
-
-    /// <summary>
-    /// Hybrid search with generative AI capabilities using hybrid vector input.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="vectors">Hybrid vector input</param>
-    /// <param name="alpha">Alpha value for hybrid search</param>
-    /// <param name="queryProperties">Properties to query</param>
-    /// <param name="fusionType">Fusion type</param>
-    /// <param name="maxVectorDistance">Maximum vector distance</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="bm25Operator">BM25 operator</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> Hybrid(
-        string? query,
-        IHybridVectorInput vectors,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            vector: vectors as Vectors,
-            nearVector: vectors as HybridNearVector,
-            nearText: vectors as HybridNearText,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Hybrid search with generative AI capabilities, grouping, and hybrid vector input.
-    /// </summary>
-    /// <param name="query">Search query</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="alpha">Alpha value for hybrid search</param>
-    /// <param name="vectors">Hybrid vector input</param>
-    /// <param name="queryProperties">Properties to query</param>
-    /// <param name="fusionType">Fusion type</param>
-    /// <param name="maxVectorDistance">Maximum vector distance</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="bm25Operator">BM25 operator</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> Hybrid(
-        string? query,
-        Models.GroupByRequest groupBy,
-        float? alpha = null,
-        IHybridVectorInput? vectors = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            vector: vectors as Vectors,
-            nearVector: vectors as HybridNearVector,
-            nearText: vectors as HybridNearText,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near object with generative AI capabilities.
-    /// </summary>
-    /// <param name="nearObject">Object ID to search near</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> NearObject(
-        Guid nearObject,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearObject(
-            _collectionClient.Name,
-            objectID: nearObject,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: null,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near object with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="nearObject">Object ID to search near</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> NearObject(
-        Guid nearObject,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearObject(
-            _collectionClient.Name,
-            objectID: nearObject,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near image with generative AI capabilities.
-    /// </summary>
-    /// <param name="nearImage">Image data</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> NearImage(
-        byte[] nearImage,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await NearMedia(
-            media: nearImage,
-            mediaType: NearMediaType.Image,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            prompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near image with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="nearImage">Image data</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> NearImage(
-        byte[] nearImage,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await NearMedia(
-            media: nearImage,
-            mediaType: NearMediaType.Image,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            prompt: prompt,
-            groupedTask: groupedTask,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near media with generative AI capabilities.
-    /// </summary>
-    /// <param name="media">Media data</param>
-    /// <param name="mediaType">Type of media</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative result</returns>
-    public async Task<GenerativeWeaviateResult> NearMedia(
-        byte[] media,
-        NearMediaType mediaType,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearMedia(
-            _collectionClient.Name,
-            media: media,
-            mediaType: mediaType,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: null,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            tenant: tenant ?? _collectionClient.Tenant,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    /// <summary>
-    /// Search near media with generative AI capabilities and grouping.
-    /// </summary>
-    /// <param name="media">Media data</param>
-    /// <param name="mediaType">Type of media</param>
-    /// <param name="groupBy">Group by configuration</param>
-    /// <param name="certainty">Certainty threshold</param>
-    /// <param name="distance">Distance threshold</param>
-    /// <param name="limit">Maximum number of results</param>
-    /// <param name="offset">Offset for pagination</param>
-    /// <param name="autoLimit">Auto-limit threshold</param>
-    /// <param name="filters">Filters to apply</param>
-    /// <param name="rerank">Rerank configuration</param>
-    /// <param name="prompt">Single prompt for generation</param>
-    /// <param name="groupedTask">Grouped prompt for generation</param>
-    /// <param name="targetVector">Target vector name</param>
-    /// <param name="tenant">Tenant name</param>
-    /// <param name="returnProperties">Properties to return</param>
-    /// <param name="returnReferences">References to return</param>
-    /// <param name="returnMetadata">Metadata to return</param>
-    /// <param name="includeVectors">Vectors to include</param>
-    /// <param name="cancellationToken">Cancellation token for the operation</param>
-    /// <returns>Generative group-by result</returns>
-    public async Task<GenerativeGroupByResult> NearMedia(
-        byte[] media,
-        NearMediaType mediaType,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        SinglePrompt? prompt = null,
-        GroupedTask? groupedTask = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await _client.GrpcClient.SearchNearMedia(
-            _collectionClient.Name,
-            media: media,
-            mediaType: mediaType,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: prompt,
-            groupedTask: groupedTask,
-            tenant: tenant ?? _collectionClient.Tenant,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        return result;
-    }
-
-    #endregion
 }
