+++ src/Weaviate.Client/QueryClient.FetchObjects.cs
@@ -1,0 +1,120 @@
+using Weaviate.Client.Models;
+
+namespace Weaviate.Client;
+
+public partial class QueryClient
+{
+    #region Objects
+    public async Task<GroupByResult> FetchObjects(
+        Models.GroupByRequest groupBy,
+        Guid? after = null,
+        uint? limit = null,
+        Filter? filters = null,
+        OneOrManyOf<Sort>? sort = null,
+        Rerank? rerank = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    )
+    {
+        return await _grpc.FetchObjects(
+            _collectionName,
+            after: after,
+            limit: limit,
+            rerank: rerank,
+            filters: filters,
+            sort: sort,
+            groupBy: groupBy,
+            tenant: _collectionClient.Tenant,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+    }
+
+    public async Task<WeaviateResult> FetchObjects(
+        Guid? after = null,
+        uint? limit = null,
+        Filter? filters = null,
+        OneOrManyOf<Sort>? sort = null,
+        Rerank? rerank = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.FetchObjects(
+            _collectionName,
+            after: after,
+            limit: limit,
+            rerank: rerank,
+            filters: filters,
+            sort: sort,
+            tenant: _collectionClient.Tenant,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+
+    public async Task<WeaviateObject?> FetchObjectByID(
+        Guid id,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    )
+    {
+        var searchReply = await _grpc.FetchObjects(
+            _collectionName,
+            returnProperties: returnProperties,
+            filters: Filter.WithID(id),
+            tenant: _collectionClient.Tenant,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+
+        WeaviateResult? result = searchReply;
+
+        WeaviateObject? firstObject = result.SingleOrDefault();
+
+        return firstObject;
+    }
+
+    public async Task<WeaviateResult> FetchObjectsByIDs(
+        HashSet<Guid> ids,
+        uint? limit = null,
+        Rerank? rerank = null,
+        Filter? filters = null,
+        OneOrManyOf<Sort>? sort = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.FetchObjects(
+            _collectionName,
+            limit: limit,
+            filters: filters != null ? Filter.WithIDs(ids) & filters : Filter.WithIDs(ids),
+            sort: sort,
+            tenant: _collectionClient.Tenant,
+            rerank: rerank,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+    #endregion
+}
