+++ src/Weaviate.Client/QueryClient.NearImage.cs
@@ -1,0 +1,80 @@
+using Weaviate.Client.Models;
+
+namespace Weaviate.Client;
+
+public partial class QueryClient
+{
+    public async Task<WeaviateResult> NearImage(
+        byte[] nearImage,
+        double? certainty = null,
+        double? distance = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    )
+    {
+        var result = await NearMedia(
+            media: nearImage,
+            mediaType: NearMediaType.Image,
+            certainty: certainty,
+            distance: distance,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            rerank: rerank,
+            targetVector: targetVector,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            cancellationToken: cancellationToken
+        );
+
+        return result;
+    }
+
+    public async Task<GroupByResult> NearImage(
+        byte[] nearImage,
+        GroupByRequest groupBy,
+        double? certainty = null,
+        double? distance = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await NearMedia(
+            media: nearImage,
+            mediaType: NearMediaType.Image,
+            certainty: certainty,
+            distance: distance,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            groupBy: groupBy,
+            rerank: rerank,
+            targetVector: targetVector,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            cancellationToken: cancellationToken
+        );
+}
