+++ src/Weaviate.Client/QueryClient.NearObject.cs
@@ -1,0 +1,85 @@
+using Weaviate.Client.Models;
+
+namespace Weaviate.Client;
+
+public partial class QueryClient
+{
+    public async Task<WeaviateResult> NearObject(
+        Guid nearObject,
+        double? certainty = null,
+        double? distance = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.SearchNearObject(
+            _collectionClient.Name,
+            objectID: nearObject,
+            certainty: certainty,
+            distance: distance,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            groupBy: null,
+            rerank: rerank,
+            singlePrompt: null,
+            groupedTask: null,
+            targetVector: targetVector,
+            tenant: _collectionClient.Tenant,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+
+    public async Task<GroupByResult> NearObject(
+        Guid nearObject,
+        GroupByRequest groupBy,
+        double? certainty = null,
+        double? distance = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.SearchNearObject(
+            _collectionClient.Name,
+            objectID: nearObject,
+            certainty: certainty,
+            distance: distance,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            groupBy: groupBy,
+            rerank: rerank,
+            singlePrompt: null,
+            groupedTask: null,
+            targetVector: targetVector,
+            tenant: _collectionClient.Tenant,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+}
