+++ src/Weaviate.Client/QueryClient.NearText.cs
@@ -1,0 +1,88 @@
+using Weaviate.Client.Models;
+
+namespace Weaviate.Client;
+
+public partial class QueryClient
+{
+    public async Task<WeaviateResult> NearText(
+        OneOrManyOf<string> text,
+        float? certainty = null,
+        float? distance = null,
+        Move? moveTo = null,
+        Move? moveAway = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.SearchNearText(
+            _collectionClient.Name,
+            text.ToArray(),
+            distance: distance,
+            certainty: certainty,
+            limit: limit,
+            moveTo: moveTo,
+            moveAway: moveAway,
+            offset: offset,
+            autoLimit: autoLimit,
+            targetVector: targetVector,
+            filters: filters,
+            tenant: _collectionClient.Tenant,
+            rerank: rerank,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+
+    public async Task<GroupByResult> NearText(
+        OneOrManyOf<string> text,
+        GroupByRequest groupBy,
+        float? certainty = null,
+        float? distance = null,
+        Move? moveTo = null,
+        Move? moveAway = null,
+        uint? limit = null,
+        uint? offset = null,
+        uint? autoLimit = null,
+        Filter? filters = null,
+        Rerank? rerank = null,
+        TargetVectors? targetVector = null,
+        OneOrManyOf<string>? returnProperties = null,
+        IList<QueryReference>? returnReferences = null,
+        MetadataQuery? returnMetadata = null,
+        VectorQuery? includeVectors = null,
+        CancellationToken cancellationToken = default
+    ) =>
+        await _grpc.SearchNearText(
+            _collectionClient.Name,
+            text.ToArray(),
+            groupBy: groupBy,
+            distance: distance,
+            certainty: certainty,
+            moveTo: moveTo,
+            moveAway: moveAway,
+            limit: limit,
+            offset: offset,
+            autoLimit: autoLimit,
+            filters: filters,
+            tenant: _collectionClient.Tenant,
+            rerank: rerank,
+            targetVector: targetVector,
+            consistencyLevel: _collectionClient.ConsistencyLevel,
+            returnProperties: returnProperties,
+            returnReferences: returnReferences,
+            returnMetadata: returnMetadata,
+            includeVectors: includeVectors,
+            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
+        );
+}
