+++ src/Weaviate.Client/QueryClient.cs
@@ -2,7 +2,7 @@ using Weaviate.Client.Models;
 
 namespace Weaviate.Client;
 
-public class QueryClient
+public partial class QueryClient
 {
     private readonly CollectionClient _collectionClient;
     private string _collectionName => _collectionClient.Name;
@@ -24,800 +24,4 @@ public class QueryClient
             _client.QueryTimeout ?? _client.DefaultTimeout ?? WeaviateDefaults.QueryTimeout;
         return TimeoutHelper.GetCancellationToken(effectiveTimeout, userToken);
     }
-
-    #region Objects
-    public async Task<GroupByResult> FetchObjects(
-        Models.GroupByRequest groupBy,
-        Guid? after = null,
-        uint? limit = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        Rerank? rerank = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        return await _grpc.FetchObjects(
-            _collectionName,
-            after: after,
-            limit: limit,
-            rerank: rerank,
-            filters: filters,
-            sort: sort,
-            groupBy: groupBy,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-    }
-
-    public async Task<WeaviateResult> FetchObjects(
-        Guid? after = null,
-        uint? limit = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        Rerank? rerank = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.FetchObjects(
-            _collectionName,
-            after: after,
-            limit: limit,
-            rerank: rerank,
-            filters: filters,
-            sort: sort,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateObject?> FetchObjectByID(
-        Guid id,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var searchReply = await _grpc.FetchObjects(
-            _collectionName,
-            returnProperties: returnProperties,
-            filters: Filter.WithID(id),
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-        WeaviateResult? result = searchReply;
-
-        WeaviateObject? firstObject = result.SingleOrDefault();
-
-        return firstObject;
-    }
-
-    public async Task<WeaviateResult> FetchObjectsByIDs(
-        HashSet<Guid> ids,
-        uint? limit = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        Filter? filters = null,
-        OneOrManyOf<Sort>? sort = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.FetchObjects(
-            _collectionName,
-            limit: limit,
-            filters: filters != null ? Filter.WithIDs(ids) & filters : Filter.WithIDs(ids),
-            sort: sort,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata?.Disable(MetadataOptions.Certainty),
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-    #endregion
-
-    #region Search
-    public async Task<WeaviateResult> NearText(
-        OneOrManyOf<string> text,
-        float? certainty = null,
-        float? distance = null,
-        Move? moveTo = null,
-        Move? moveAway = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        string? tenant = null,
-        TargetVectors? targetVector = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearText(
-            _collectionClient.Name,
-            text.ToArray(),
-            distance: distance,
-            certainty: certainty,
-            limit: limit,
-            moveTo: moveTo,
-            moveAway: moveAway,
-            offset: offset,
-            autoLimit: autoLimit,
-            targetVector: targetVector,
-            filters: filters,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> NearText(
-        OneOrManyOf<string> text,
-        GroupByRequest groupBy,
-        float? certainty = null,
-        float? distance = null,
-        Move? moveTo = null,
-        Move? moveAway = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearText(
-            _collectionClient.Name,
-            text.ToArray(),
-            groupBy: groupBy,
-            distance: distance,
-            certainty: certainty,
-            moveTo: moveTo,
-            moveAway: moveAway,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            tenant: _collectionClient.Tenant,
-            rerank: rerank,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> NearVector(
-        Vectors vector,
-        Filter? filters = null,
-        float? certainty = null,
-        float? distance = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearVector(
-            _collectionClient.Name,
-            vector,
-            distance: distance,
-            certainty: certainty,
-            offset: offset,
-            autoLimit: autoLimit,
-            limit: limit,
-            targetVector: targetVector,
-            filters: filters,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> NearVector(
-        Vectors vector,
-        GroupByRequest groupBy,
-        Filter? filters = null,
-        float? distance = null,
-        float? certainty = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        Rerank? rerank = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearVector(
-            _collectionClient.Name,
-            vector,
-            groupBy,
-            filters: filters,
-            distance: distance,
-            certainty: certainty,
-            offset: offset,
-            autoLimit: autoLimit,
-            limit: limit,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            rerank: rerank,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> BM25(
-        string query,
-        GroupByRequest groupBy,
-        string[]? searchFields = null,
-        Filter? filters = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? searchOperator = null,
-        Rerank? rerank = null,
-        Guid? after = null,
-        string? tenant = null,
-        ConsistencyLevels? consistencyLevel = null,
-        OneOrManyOf<string>? returnProperties = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        IList<QueryReference>? returnReferences = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchBM25(
-            _collectionClient.Name,
-            query: query,
-            searchFields: searchFields,
-            filters: filters,
-            autoLimit: autoLimit,
-            limit: limit,
-            offset: offset,
-            searchOperator: searchOperator,
-            groupBy: groupBy,
-            rerank: rerank,
-            after: after,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: consistencyLevel ?? _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnReferences: returnReferences,
-            returnProperties: returnProperties,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> BM25(
-        string query,
-        string[]? searchFields = null,
-        Filter? filters = null,
-        uint? autoLimit = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? searchOperator = null,
-        Rerank? rerank = null,
-        Guid? after = null,
-        string? tenant = null,
-        ConsistencyLevels? consistencyLevel = null,
-        OneOrManyOf<string>? returnProperties = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        IList<QueryReference>? returnReferences = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchBM25(
-            _collectionClient.Name,
-            query: query,
-            searchFields: searchFields,
-            filters: filters,
-            autoLimit: autoLimit,
-            limit: limit,
-            offset: offset,
-            searchOperator: searchOperator,
-            groupBy: null,
-            rerank: rerank,
-            after: after,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: consistencyLevel ?? _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnReferences: returnReferences,
-            returnProperties: returnProperties,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> Hybrid(
-        string? query,
-        Vectors vectors,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            vector: vectors,
-            nearVector: null,
-            nearText: null,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> Hybrid(
-        string? query,
-        IHybridVectorInput? vectors = null,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            alpha: alpha,
-            vector: vectors is Vector v ? Vectors.Create(v) : vectors as Vectors,
-            nearVector: vectors as HybridNearVector,
-            nearText: vectors as HybridNearText,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> Hybrid(
-        string? query,
-        Models.GroupByRequest groupBy,
-        Vectors vectors,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            vector: vectors,
-            nearVector: null,
-            nearText: null,
-            alpha: alpha,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> Hybrid(
-        string? query,
-        Models.GroupByRequest groupBy,
-        IHybridVectorInput? vectors = null,
-        float? alpha = null,
-        string[]? queryProperties = null,
-        HybridFusion? fusionType = null,
-        float? maxVectorDistance = null,
-        uint? limit = null,
-        uint? offset = null,
-        BM25Operator? bm25Operator = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchHybrid(
-            _collectionClient.Name,
-            query: query,
-            vector: vectors is Vector v ? Vectors.Create(v) : vectors as Vectors,
-            nearVector: vectors as HybridNearVector,
-            nearText: vectors as HybridNearText,
-            alpha: alpha,
-            queryProperties: queryProperties,
-            fusionType: fusionType,
-            maxVectorDistance: maxVectorDistance,
-            limit: limit,
-            offset: offset,
-            bm25Operator: bm25Operator,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> NearObject(
-        Guid nearObject,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearObject(
-            _collectionClient.Name,
-            objectID: nearObject,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: null,
-            rerank: rerank,
-            singlePrompt: null,
-            groupedTask: null,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> NearObject(
-        Guid nearObject,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearObject(
-            _collectionClient.Name,
-            objectID: nearObject,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: null,
-            groupedTask: null,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<WeaviateResult> NearImage(
-        byte[] nearImage,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var result = await NearMedia(
-            media: nearImage,
-            mediaType: NearMediaType.Image,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: cancellationToken
-        );
-
-        return result;
-    }
-
-    public async Task<GroupByResult> NearImage(
-        byte[] nearImage,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await NearMedia(
-            media: nearImage,
-            mediaType: NearMediaType.Image,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            targetVector: targetVector,
-            tenant: tenant ?? _collectionClient.Tenant,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: cancellationToken
-        );
-
-    public async Task<WeaviateResult> NearMedia(
-        byte[] media,
-        NearMediaType mediaType,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearMedia(
-            _collectionClient.Name,
-            media: media,
-            mediaType: mediaType,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: null,
-            rerank: rerank,
-            singlePrompt: null,
-            groupedTask: null,
-            tenant: tenant ?? _collectionClient.Tenant,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    public async Task<GroupByResult> NearMedia(
-        byte[] media,
-        NearMediaType mediaType,
-        GroupByRequest groupBy,
-        double? certainty = null,
-        double? distance = null,
-        uint? limit = null,
-        uint? offset = null,
-        uint? autoLimit = null,
-        Filter? filters = null,
-        Rerank? rerank = null,
-        TargetVectors? targetVector = null,
-        string? tenant = null,
-        OneOrManyOf<string>? returnProperties = null,
-        IList<QueryReference>? returnReferences = null,
-        MetadataQuery? returnMetadata = null,
-        VectorQuery? includeVectors = null,
-        CancellationToken cancellationToken = default
-    ) =>
-        await _grpc.SearchNearMedia(
-            _collectionClient.Name,
-            media: media,
-            mediaType: mediaType,
-            certainty: certainty,
-            distance: distance,
-            limit: limit,
-            offset: offset,
-            autoLimit: autoLimit,
-            filters: filters,
-            groupBy: groupBy,
-            rerank: rerank,
-            singlePrompt: null,
-            groupedTask: null,
-            tenant: tenant ?? _collectionClient.Tenant,
-            targetVector: targetVector,
-            consistencyLevel: _collectionClient.ConsistencyLevel,
-            returnMetadata: returnMetadata,
-            includeVectors: includeVectors,
-            returnProperties: returnProperties,
-            returnReferences: returnReferences,
-            cancellationToken: CreateTimeoutCancellationToken(cancellationToken)
-        );
-
-    #endregion
 }
